chore: move MCPIcon to correct location, remove duplicate
